# shared/openapi/mesob-api.yaml
openapi: "3.1.0"
info:
  title: Mesob Wallet API
  version: "1.0.0"
  description: |
    Unified public API for the Mesob Wallet platform.
    All requests go through the API Gateway at port 8000.

    ## Authentication
    - Mobile/USSD users authenticate with **MSISDN + PIN** via `/v1/identity/login`
    - Access tokens are RS256 JWTs valid for **15 minutes**
    - Refresh tokens are 7-day opaque tokens returned as an `httpOnly` cookie
    - Staff users (ADMIN, BRANCH_MANAGER, SUPER_ADMIN) use the same auth flow

    ## Demo Accounts
    | Role | MSISDN | PIN |
    |------|--------|-----|
    | SUPER_ADMIN | +251911000001 | 111111 |
    | ADMIN | +251911000002 | 111111 |
    | BRANCH_MANAGER | +251911000003 | 111111 |
    | AGENT | +251911000004 | 111111 |
    | USER (KYC Tier 2) | +251911000005 | 111111 |
    | USER (KYC Tier 0) | +251911000006 | 111111 |

    ## Amounts
    All monetary values are in **ETB minor units (cents)**.
    Example: 100.00 ETB = `10000`

    ## Response Envelope
    All responses follow:
    ```json
    {
      "data": { ... },
      "meta": { "request_id": "...", "trace_id": "..." }
    }
    ```
    Errors:
    ```json
    {
      "error": { "code": "DOMAIN_ERROR_CODE", "message": "human message" },
      "meta": { ... }
    }
    ```
  contact:
    name: Mesob Engineering
    email: engineering@mesobwallet.et
servers:
  - url: http://localhost:8000/v1
    description: Local development
  - url: https://api.mesobwallet.et/v1
...
paths:
...
  /identity/verify-otp:
    post:
      tags: [Auth]
      summary: Verify OTP — get challenge token for PIN setup
      operationId: verifyOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [registration_id, otp]
              properties:
                registration_id:
                  type: string
                  example: "reg_01J3ABCDEF"
                otp:
                  type: string
                  minLength: 6
                  maxLength: 6
                  example: "482910"
      responses:
        "200":
          description: OTP verified — challenge token issued for PIN creation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
              example:
                data:
                  challenge_token: "chall_01J3XYZPQR"
                  expires_in: 600
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /identity/token/refresh:
    post:
      tags: [Auth]
      summary: Rotate access token using refresh token cookie
      operationId: tokenRefresh
      description: |
        Reads the `rt` httpOnly cookie (or `X-Refresh-Token` header for mobile).
        Implements token-family rotation — using a revoked token invalidates the whole family.
      parameters:
        - in: header
          name: X-Refresh-Token
          schema:
            type: string
          description: Mobile alternative to the `rt` cookie
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
              example:
                data:
                  access_token: "eyJhbGciOiJSUzI1NiJ9..."
                  refresh_token: "rt_new_opaque_token"
                  expires_in: 900
                  role: "USER"
                  kyc_tier: 2
                  user_id: "01924b6f-1234-7abc-8def-000000000005"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /identity/me/language:
    patch:
      tags: [Identity]
      summary: Update preferred language
      operationId: patchLanguage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lang]
              properties:
                lang:
                  type: string
                  enum: [am, om, ti, en]
      responses:
        "200":
          description: Language updated
        "401":
          $ref: "#/components/responses/Unauthorized"
  /identity/pin/change:
    post:
      tags: [Identity]
      summary: Change PIN (requires current PIN)
      operationId: changePin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_pin, new_pin]
              properties:
                current_pin:
                  type: string
                  minLength: 4
                  maxLength: 6
                new_pin:
                  type: string
                  minLength: 4
                  maxLength: 6
      responses:
        "200":
          description: PIN changed, all sessions revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /identity/kyc/upgrade:
    post:
      tags: [Identity]
      summary: Request KYC tier upgrade
      operationId: kycUpgrade
      security:
        - bearerAuth: []
      description: |
        Submits KYC upgrade request. Triggers NID verification via adapter-hub.
        - Tier 1: requires FAN (Fayda ID number) + name + DOB
        - Tier 2: requires Tier 1 + address proof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_tier, fan]
              properties:
                target_tier:
                  type: integer
                  enum: [1, 2]
                fan:
                  type: string
                  description: Fayda Account Number (Ethiopian National ID)
                  example: "FAN-1234567890"
                full_name:
                  type: string
                  example: "Abebe Girma"
                date_of_birth:
                  type: string
                  format: date
                  example: "1990-05-15"
      responses:
        "202":
          description: KYC review submitted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  # ── PAYMENTS ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT access token. Obtain via /identity/login or /identity/token/refresh.
...
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        data:
          description: Response payload (schema varies by endpoint)
        meta:
          $ref: "#/components/schemas/MetaBlock"
      required: [data]
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "INSUFFICIENT_BALANCE"
            message:
              type: string
              description: Human-readable message
              example: "Your balance is too low for this transaction"
        meta:
          $ref: "#/components/schemas/MetaBlock"
      required: [error]
    MetaBlock:
      type: object
      properties:
        request_id:
          type: string
          example: "req_01J3ABCDEF"
        trace_id:
          type: string
          example: "trace_01J3ABCDEF"
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Access token missing or expired"
...
    UnprocessableEntity:
      description: Domain validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: "INVALID_MSISDN"
              message: "Phone number must be in E.164 format"